Moved for loop tests into vmForLoop_test.go
